+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version: 3.2.1
 
 ENHANCEMENTS:
 
+* Improvement done in the way the gallery files are copied to the output directory.
 * Improvement done in before_scenario hook add robustness on the execution logic.